Add support for specifying a `beforeCommit` hook.
If you supply a function as the `beforeCommit` property
in your config/release.js, it will be invoked with a reference
to the `project` and the `versions` object. The purpose of the
hook is to allow a project to do any other updates to files
that may need to be updated to the current version.

e.g.

```
var path = require('path');
var xmlpoke = require('xmlpoke');

module.exports = {
  beforeCommit: function(project, versions){
    xmlpoke(path.join(project.root, 'cordova/config.xml'), function(xml) {
      xml.addNamespace('w', 'http://www.w3.org/ns/widgets');
      xml.set('w:widget/@version', versions.next);
    });
  }
}
```