Fix bug where config file was not being loaded.
An incorrect file existence check resulted in the config file never
actually being loaded.

In testing this functionality, we also realized that the goal of having
explicit CLI options take precedence over the config file is not
currently possible. We can’t distinguish between explicitly specified
CLI options and their default values because ember-cli pre-applies the
default values and passes us the merged result.

In light of this and in order to provide a consistent experience, this
commit updates the behavior and documentation to make options specified
via the config file take precedence over explicit CLI options and their
default values.