Removed all but 'beforeCommit' and 'afterPush' hooks
The vast majority of use cases for lifecycle hooks are achieved with
`beforeCommit`, and the only other apparent use cases happen after all
operations are complete.

In the interest of not introducing unnecessary APIs, this commit removes
all hooks in between git operations. Reverting this commit will restore
them should the need arise.